// app/shared/dto/institution/institution.dto.ts
export class InstitutionDto {
    institutionId: string;
    shortName: string;
    firmenbuchnumber: string;
    uidNumber: string;
    zipCode: string;
    street: string;
    country: string;
    city: string;
    email: string;
    phoneFixed: string;
    phoneMobile: string;
    iban: string;
    bic: string;
    bankName: string;
    institutionsName: string;
    emailSenderName: string;
    emailReplayTo: string;
    emailAdmin: string;
    languages: string;
    isActive: boolean;
    webSite: string;
    currencySign: string;
    newsletterLink: string;
    isStripeActive: boolean;
    stripePublishableKey: string;
}